requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
lancedb>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
pydantic>=1.10.0
python-dotenv>=1.0.0
requests>=2.28.0

# ML dependencies
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
torch>=2.0.0
scipy>=1.10.0

# Monitoring and utilities
prometheus-client>=0.16.0
python-json-logger>=2.0.7
httpx>=0.24.0
aiofiles>=23.1.0
tenacity>=8.2.0

# Storage and serialization
aiofiles>=0.8.0
pyarrow>=11.0.0
joblib>=1.2.0

# Optional dependencies (enabled via feature flags)
# Uncomment if you need them
# mlflow>=2.3.0  # For experiment tracking
# boto3>=1.26.0  # For S3 storage
# psycopg2-binary>=2.9.0  # For PostgreSQL connection
# pymongo>=4.3.0  # For MongoDB connection

sqlalchemy
